+++ R/i_concat.R
@@ -1,10 +1,10 @@
 #' @rdname i_chain
 #' @param obj an iterable.
-#' @description `icollapse(obj)` takes an iterable that returns
+#' @description `i_concat(obj)` takes an iterable that returns
 #'   iterables, and chains together all inner values of iterables into
 #'   one iterator. Analogous to `unlist(recursive=FALSE)`.
 #' @export
-icollapse <- function(obj, ...) {
+i_concat <- function(obj, ...) {
   obj <- iteror(obj, ...)
   current <- NULL